Added debugging
