+++ frontend/js/tile.js
@@ -101,3 +101,47 @@ document.addEventListener('DOMContentLoaded', function() {
   });
   setBackgroundColor();
 });
+
+export async function updateButtonTexts(userId) {
+  const bgButton = document.querySelectorAll('#bgBlue, #bgBrown, #bgCrimson, #bgGreen, #bgGrey, #bgOrange, #bgPink, #bgRed');
+  for (const button of bgButton) {
+      let color;
+      switch (button.id) {
+      case 'bgBlue':
+          color = '#00a3e9';
+          break;
+      case 'bgBrown':
+          color = '#b97b56';
+          break;
+      case 'bgCrimson':
+          color = '#7b0103';
+          break;
+      case 'bgGreen':
+          color = '#22b14c';
+          break;
+      case 'bgGrey':
+          color = '#7f7f7f';
+          break;
+      case 'bgOrange':
+          color = '#fc6a03';
+          break;
+      case 'bgPink':
+          color = '#eb3780';
+          break;
+      case 'bgRed':
+          color = '#ed1d25';
+          break;
+      }
+      if (color) {
+          const isOwned = await isBackgroundOwned(userId, color);
+          console.log(`Color: ${color}, Owned: ${isOwned}`);
+          const buttonText = isOwned ? "Change" : "Add";
+          const buttonTag = document.getElementById(button.id);
+          if (buttonTag) {
+            buttonTag.innerHTML = ` ${buttonText} `;
+          } else {
+              console.error(`Button with ID ${button.id} not found`);
+          }
+      }
+  }
+}
